feat: Phase 2.5 UX 개선 - 즐겨찾기, DnD 정렬, 수량입력 개선, 검색, 다중선택 레시피
- 즐겨찾는 냉장고: 별 토글(PATCH favorite), 즐겨찾기 상위 정렬
- 냉장고/보관함 드래그앤드롭 순서 변경 (@dnd-kit/sortable)
- 수량 입력 min 제약 제거, 제출 시 검증 (0.5 자연스럽게 입력 가능)
- 대시보드 식재료 검색 (debounced ILIKE)
- 다중 냉장고 선택 레시피 추천
- 냉장고 카드 D-day 배지, 단위 입력 커스텀/목록 토글
- 수량 int → float 전환
- Alembic 마이그레이션 (is_favorite, display_order)

+++ backend/app/models/ingredient.py
@@ -1,6 +1,6 @@
 from datetime import date
 
-from sqlalchemy import Date, ForeignKey, Index, Integer, String
+from sqlalchemy import Date, Float, ForeignKey, Index, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database.base import Base
@@ -14,7 +14,7 @@ class Ingredient(Base):
         ForeignKey("storage_boxes.id", ondelete="CASCADE"), nullable=False
     )
     name: Mapped[str] = mapped_column(String(100), nullable=False)
-    quantity: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
+    quantity: Mapped[float] = mapped_column(Float, nullable=False, default=1)
     unit: Mapped[str] = mapped_column(String(20), nullable=False, default="개")
     expiry_date: Mapped[date | None] = mapped_column(Date, nullable=True)
